Fix crash in libecho.c on win64 build
Return value of _findfirst is intptr_t; see
https://msdn.microsoft.com/en-us/library/zyzxfzac.aspx

[ghudson@mit.edu: also fix plugins.c]

ticket: 8230

+++ src/util/support/plugins.c
@@ -96,7 +96,7 @@ struct dirent {
 };
 
 typedef struct {
-    long handle;                /* _findfirst/_findnext handle */
+    intptr_t handle;            /* _findfirst/_findnext handle */
     short offset;               /* offset into directory */
     short finished;             /* 1 if there are not more files */
     struct _finddata_t fileinfo;/* from _findfirst/_findnext */
@@ -108,7 +108,7 @@ DIR * opendir(const char *dir)
 {
     DIR *dp;
     char *filespec;
-    long handle;
+    intptr_t handle;
     int index;
 
     filespec = malloc(strlen(dir) + 2 + 1);
